requirements.txt
anyio==4.8.0
asgiref==3.8.1
certifi==2024.12.14
Django==5.1.4
et_xmlfile==2.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
isort==5.13.2
openpyxl==3.1.5
python-dotenv==1.0.1
python-telegram-bot==21.10
setuptools==75.8.0
sniffio==1.3.1
sqlparse==0.5.3
wheel==0.45.1
